ingestion/requirements.txt
faker==37.1.0
google-cloud-storage==2.19.0
pyarrow==19.0.0
requests==2.32.3